chore: Move utils/cloudprofile to pkg/utils/gardener/cloudprofile
